Add a new option to WebFrontAuthOptions : SlidingExpirationRefreshLimit:
If set to true, it will behave kind of like dotnet cookie.
It means that the expiration time will be refreshed only when it processes a request which is more than halfway through the expiration window.
Since ExpirationTimeSpan and SlidingExpirationTime exists, this code will be ignored and throw a warning whenever those 2 values are not equals.

This commit has been issued to solve antiforgery issues.